Timeout on pause attempt to prevent multiple executions
A long query with AccessShareLock on master will be replicated,
and will cause an internal loop when invoking pg_back on standby.
To prevent multiple executions of pg_back on daily routine, this
introduces a new option PGBK_PAUSE_TIMEOUT (-T) that waits an amount
of time in seconds before dying.

The pause/resume replay logic has been simplified.