allow postgresql URIs as connection strings
Add parsing and detection of postgresql:// URIs. Parsing is done with
the net/url module from the Go standard library with added support for a
comma separated list of hosts. To allow extracting and modifying connection
information, we add new ConnInfo struct with some methods to update
information, such as dbname, that work in concurrent goroutines. Instead
of carrying host, port, etc around, a ConnInfo is used, so we can parse
the input URI and fail earlier with a better message than having the
building of the pg_dumpall fail.