encrypt and decrypt files with age
Add options (cli and conf) to encrypt and decrypt the files produced by
a run with age. Age is a file encryption tool that supports keys and
passphrases. To keep things simple, we only support passphrase. The
passphrase can be on the command line, in the config file or set by the
PGBK_PASSPHRASE environment variable.