check and adapt to the version of pg_dump and pg_dumpall
When pg_dump/all is older than 9.3 pass the connection string as
environment variables instead of using -d which does not exist. Also,
fail when the user asks for the directory format and pg_dump does not
know it (<9.1). For other unknown parameters (-j and -B) ignore them
with a warning if pg_dump is too old.