fix purge based on a number to keep
Purge of dumps while keeping a number of files did not make a difference
between dumps and other files like checksum, encrypted files or createdb
DDL script. So now, the files are first grouped by date to identify if
they were created by the run and the number of groups is used to compute
the number of dumps to keep.