AGE encryption and decryption with asymmetric keys
This allows pg_back to encrypt the backups using a public key only. This
has the advantage of not leaking the decryption key for other backups
if an attacker gains access to the backup system.

This helps with the Kubernetes usecase where the secret may be
long-lived and we want to minimize attack surface when some part of the
system is compromised.