add options to avoid dumping hash passwords with pg_dumpall
Some providers do not allow superuser access so pg_dumpall can extract
hashes of passwords. Add the --with-role-passwords and the
dump_role_passwords config parameter (defaults to true), the cli option
can be used to overwrite the value of the configuration file,
--without-role-passwords disable dumping of roles by making pg_back add
the --no-role-passwords to pg_dumpall. This is possible pg_dumpall from
PostgreSQL 10, older version makes pg_back exit with an error.