ensure usability without superuser privileges
Check if the connection user is a superuser and when not, do not extract
configuration from files as it is not possible. We prefer not to dump
from pg_settings as there is no way to know if the configuration as been
change at a higher level than the configuration file without more
privileges. Also query pg_roles instead of pg_authid and force dump of
roles without passwords with pg_dumpall.