Add an option to configure permission on dump
Add an option / configuration item to define the permissions for the
dumps and hash files created by pg_back. The 0600 permissions mode is
defined by default to ensure compatibility with the old (non configurable)
approach.

Parent directory when using 'd' for the dump format still use the 0700
mode, but may be the configured mode should be applied on those
situations.