Add an option to configure permission on dump
We add an option / configuration item to define the permissions for the
dumps created with pg_back. We reuse / configure the 0600 permissions by
default to ensure some compatibility with the old (non configurable)
approach.

We probably need to discuss and improve what we should do for the parent
directory when using 'd' for the dump format. Currently we still apply
the 0700 mode in such situation, but maybe we should define a specific
option or reuse the one we add right now.